+++ inertia/inertia.py
@@ -1,6 +1,5 @@
 import logging
 import os
-import warnings
 
 from fastapi import Depends, Request, Response, status
 from fastapi.responses import JSONResponse, HTMLResponse
@@ -21,15 +20,14 @@ import json
 from pydantic import BaseModel
 from starlette.responses import RedirectResponse
 
-from .config import InertiaConfig, _read_manifest_file
+
+from .templating import InertiaExtension
+from .config import InertiaConfig
+from .utils import InertiaContext, _read_manifest_file
 from .exceptions import InertiaVersionConflictException
 from .utils import LazyProp
 from dataclasses import dataclass
 
-try:
-    import requests
-except (ModuleNotFoundError, ImportError):
-    requests = None  # type: ignore
 
 try:
     import httpx
@@ -72,6 +70,7 @@ class Inertia:
     _request: Request
     _component: str
     _props: dict[str, Any]
+    _config: InertiaConfig
     _inertia_files: InertiaFiles
     _client: Union["httpx.AsyncClient", None]
 
@@ -169,17 +168,14 @@ class Inertia:
             else {}
         )
 
-    def _assert_a_request_package_is_installed(self) -> None:
-        if not httpx and not requests:
-            raise ImportError(
-                "You need to install either requests or httpx to use Inertia in SSR mode"
-            )
-        if requests:
-            warnings.warn(
-                "requests is deprecated: Please use httpx instead. It will be removed in 1.0.0",
-                DeprecationWarning,
-                stacklevel=2,
-            )
+    @staticmethod
+    def _assert_httpx_is_installed() -> None:
+        """
+        Assert that httpx is installed
+        :raises ImportError: If httpx is not installed
+        """
+        if not httpx:
+            raise ImportError("You need to install httpx to use Inertia in SSR mode")
 
     def _set_inertia_files(self) -> None:
         """
@@ -249,57 +245,23 @@ class Inertia:
 
         return self._deep_transform_callables(_props)
 
-    def _get_html_content(self, head: str, body: str) -> str:
-        """
-        Get the HTML content for the response
-        :param head: The content for the head tag
-        :param body: The content for the body tag
-        :return: The HTML content
-        """
-        css_links = (
-            "\n".join(
-                [
-                    f'<link rel="stylesheet" href="{url}">'
-                    for url in self._inertia_files.css_file_urls
-                ]
-            )
-            if len(self._inertia_files.css_file_urls) > 0
-            else ""
-        )
-
-        return f"""
-                   <!DOCTYPE html>
-                   <html>
-                       <head>
-                            <meta charset="UTF-8">
-                            <meta name="viewport" content="width=device-width, initial-scale=1.0">
-                            {head}
-                            {css_links}
-                        </head>
-                        <body>
-                            {body}
-                            <script type="module" src="{self._inertia_files.js_file_url}"></script>
-                       </body>
-                   </html>
-                   """
-
     async def _render_ssr(self) -> HTMLResponse:
         """
         Render the page using SSR, calling the Inertia SSR server.
         :return: The HTML response
         """
-        self._assert_a_request_package_is_installed()
+        self._assert_httpx_is_installed()
         data = json.dumps(self._get_page_data(), cls=self._config.json_encoder)
         request_kwargs: Dict[str, Any] = {
             "url": f"{self._config.ssr_url}/render",
             "json": data,
             "headers": {"Content-Type": "application/json"},
         }
-        response: Union["httpx._models.Response", "requests.Response"]
-        if self._client is not None:
-            response = await self._client.post(**request_kwargs)
-        else:
-            response = requests.post(**request_kwargs)
+
+        if self._client is None:
+            raise ValueError("httpx client is not available")
+
+        response = await self._client.post(**request_kwargs)
 
         response.raise_for_status()
         response_json = response.json()
@@ -308,9 +270,21 @@ class Inertia:
         displayable_head = "\n".join(head)
         body = response_json["body"]
 
-        html_content = self._get_html_content(displayable_head, body)
-
-        return HTMLResponse(content=html_content, status_code=200)
+        return self._config.templates.TemplateResponse(
+            name=self._config.root_template_filename,
+            request=self._request,
+            context={
+                "inertia": InertiaContext(
+                    environment=self._config.environment,
+                    dev_url=self._config.dev_url,
+                    is_ssr=True,
+                    ssr_head=displayable_head,
+                    ssr_body=body,
+                    js=self._inertia_files.js_file_url,
+                    css=self._inertia_files.css_file_urls,
+                ),
+            },
+        )
 
     def _render_json(self) -> JSONResponse:
         """
@@ -385,7 +359,7 @@ class Inertia:
         If an error occurs, it will fall back to server-side template rendering
         :param component: The component name to render
         :param props: The props to pass to the component
-        :return: HTMLResponse or JSONResponse
+        :return: InertiaResponse
         """
         if self._config.use_flash_messages:
             self._props.update(
@@ -415,19 +389,27 @@ class Inertia:
         page_json = json.dumps(
             json.dumps(self._get_page_data(), cls=self._config.json_encoder)
         )
-        body = f"<div id='app' data-page='{page_json}'></div>"
-        html_content = self._get_html_content("", body)
-
-        return HTMLResponse(content=html_content)
+        return self._config.templates.TemplateResponse(
+            name=self._config.root_template_filename,
+            request=self._request,
+            context={
+                "inertia": InertiaContext(
+                    environment=self._config.environment,
+                    dev_url=self._config.dev_url,
+                    is_ssr=False,
+                    data=page_json,
+                    js=self._inertia_files.js_file_url,
+                    css=self._inertia_files.css_file_urls,
+                ),
+            },
+        )
 
 
 async def get_httpx_client() -> AsyncGenerator[Union[None, "httpx.AsyncClient"], None]:
-    try:
-        async with httpx.AsyncClient() as client:
-            yield client
-    except Exception as e:
-        logger.error(f"An error occurred in creating the HTTPX client: {e}")
+    if not httpx:
         yield None
+    async with httpx.AsyncClient() as client:
+        yield client
 
 
 HttpxClientDep = Annotated[Union["httpx.AsyncClient", None], Depends(get_httpx_client)]
@@ -442,9 +424,12 @@ def inertia_dependency_factory(
     :return: Dependency
     """
 
+    config_.templates.env.add_extension(InertiaExtension)
+
     def inertia_dependency(request: Request, client: HttpxClientDep) -> Inertia:
         """
         Dependency for Inertia
+        :param client: HttpxClientDep
         :param request: FastAPI Request object
         :return: Inertia object
         """