+++ inertia/tests/test_back_helper.py
@@ -5,10 +5,13 @@ from starlette.responses import RedirectResponse
 from starlette.testclient import TestClient
 
 from inertia import Inertia, InertiaResponse, inertia_dependency_factory, InertiaConfig
+from .utils import templates
 
 
 app = FastAPI()
-InertiaDep = Annotated[Inertia, Depends(inertia_dependency_factory(InertiaConfig()))]
+InertiaDep = Annotated[
+    Inertia, Depends(inertia_dependency_factory(InertiaConfig(templates=templates)))
+]
 
 
 PROPS = {